chore: refactor gallery area
