+++ apps/web/components/ZoomableImage/types.ts
@@ -1,0 +1,9 @@
+import type { ImagesData } from '@plentymarkets/shop-api';
+
+export type ZoomableImageProps = {
+  images: ImagesData[];
+  image: ImagesData;
+  index: number;
+  activeIndex: number;
+  isFirstImage: boolean;
+};
